backend/requirements.txt
asgiref
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
django-bootstrap5
django-crispy-forms
django-debug-toolbar
django-environ
dotenv

